% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurchaserConfirm.R
\name{PurchaserConfirmUI}
\alias{PurchaserConfirmUI}
\title{采购经理确认任务生成生成器界面}
\usage{
PurchaserConfirmUI(
  tabTitle = "采购员确认",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = PurchaserConfirmUI_left,
  func_right = PurchaserConfirmUI_right,
  func_bottom = PurchaserConfirmUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
采购经理确认任务生成生成器界面
}
\examples{
PurchaserConfirmUI()
}
